Added RadixIPLookup elements which vary in their branching factor and the way they store member fields like "_nchildren", "_n", etc. Each of the RadixIPLookup[1-7] elements differ in either the branching factor or the memory which it uses.
The contents in the perf directory are used to test the performance of
each one of these modified RadixIPLookup elements on a randomly generated
input of 5 million IP addresses. The routing table used in these tests
is taken from the 167k click regression test.

The files in the test directory have been modified/added which test the
correctness of the new RadixIPLookup elements. They are essentially the
same tests which were present for the RadixIPLookup element.